NPL/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asttokens==2.4.1
biopython==1.85
blis==1.3.0
breadability==0.1.20
catalogue==2.0.10
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.6
decorator==5.1.1
dnspython==2.8.0
docopt==0.6.2
email-validator==2.3.0
executing==2.1.0
fastapi==0.116.1
fonttools==4.59.2
h11==0.16.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.13
nest-asyncio==1.6.0
nltk==3.9.2
numpy==2.3.3
packaging==25.0
pandas==2.3.2
parso==0.8.4
pillow==11.3.0
platformdirs==4.4.0
preshed==3.0.10
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
pycountry==24.6.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.2.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.3
sumy==0.11.0
thinc==8.3.6
threadpoolctl==3.6.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wheel==0.45.1
wordcloud==1.9.4
wrapt==1.17.3
